refactor: integrate signup step with form wizard
