+++ components/organisms/EventForm.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import React, { useEffect, useState, useMemo } from 'react'
+import React, { useEffect, useState, useMemo, useRef } from 'react'
 import Image from 'next/image'
 import { useTenant } from '@/lib/context/TenantContext'
 import { useToast } from '@/lib/context/ToastContext'
@@ -8,18 +8,14 @@ import { useRouter } from 'next/navigation'
 import { useAuthContext } from '@/lib/context/AuthContext'
 import createPocketBase from '@/lib/pocketbase'
 import { getAuthHeaders } from '@/lib/authHeaders'
-import { fetchCep } from '@/utils/cep'
-import { isValidCPF, isValidEmail } from '@/utils/validators'
 import FormWizard from './FormWizard'
 import LoadingOverlay from './LoadingOverlay'
 import InscricoesTable from '@/app/cliente/components/InscricoesTable'
 import type { Inscricao } from '@/types'
-import {
-  FormField,
-  InputWithMask,
-  TextField,
-  PasswordField,
-} from '@/components'
+import { FormField } from '@/components'
+import CreateUserForm, {
+  type CreateUserFormHandle,
+} from '../templates/CreateUserForm'
 
 interface Produto {
   id: string
@@ -61,31 +57,18 @@ export default function EventForm({ eventoId, liderId }: EventFormProps) {
   const [produtos, setProdutos] = useState<Produto[]>([])
   const [cobraInscricao, setCobraInscricao] = useState(false)
   const [form, setForm] = useState({
-    nome: '',
-    email: '',
-    telefone: '',
-    cpf: '',
-    data_nascimento: '',
     genero: '',
-    cep: '',
-    endereco: '',
-    bairro: '',
-    estado: '',
-    cidade: '',
-    numero: '',
     campoId: '',
     produtoId: '',
     tamanho: '',
-    role: 'usuario',
-    password: '',
-    passwordConfirm: '',
     paymentMethod: 'pix',
   })
   const [loading, setLoading] = useState(false)
   const [fetching, setFetching] = useState(true)
-  const [fieldErrors, setFieldErrors] = useState<{ cpf?: string; email?: string }>({})
   const [pendentes, setPendentes] = useState<Inscricao[]>([])
   const [checouPendentes, setChecouPendentes] = useState(false)
+  const [signupDone, setSignupDone] = useState(isLoggedIn)
+  const createUserRef = useRef<CreateUserFormHandle>(null)
 
   useEffect(() => {
     async function fetchData() {
@@ -159,29 +142,6 @@ export default function EventForm({ eventoId, liderId }: EventFormProps) {
   }, [eventoId, liderId, pb])
 
   useEffect(() => {
-    if (isLoggedIn && user) {
-      const nasc = String(user.data_nascimento ?? '')
-      setForm((prev) => ({
-        ...prev,
-        nome: user.nome || '',
-        email: user.email || '',
-        telefone: user.telefone || '',
-        cpf: user.cpf || '',
-        data_nascimento: nasc ? nasc.split(' ')[0] : '',
-        genero: (user as Record<string, string>).genero || '',
-        cep: user.cep || '',
-        endereco: user.endereco || '',
-        bairro: user.bairro || '',
-        estado: user.estado || '',
-        cidade: user.cidade || '',
-        numero: user.numero || '',
-        role: 'usuario',
-        campoId: (user.campo as string) || prev.campoId,
-      }))
-    }
-  }, [isLoggedIn, user])
-
-  useEffect(() => {
     async function verificarPendentes() {
       if (!isLoggedIn || !user) {
         setChecouPendentes(true)
@@ -212,145 +172,49 @@ export default function EventForm({ eventoId, liderId }: EventFormProps) {
     }
     verificarPendentes()
   }, [isLoggedIn, user, eventoId, pb])
-
-  useEffect(() => {
-    async function lookup() {
-      const data = await fetchCep(form.cep).catch(() => null)
-      if (!data) {
-        showError('CEP n\u00e3o encontrado.')
-        setForm((prev) => ({
-          ...prev,
-          endereco: '',
-          bairro: '',
-          cidade: '',
-          estado: '',
-        }))
-        return
-      }
-      setForm((prev) => ({
-        ...prev,
-        endereco: data.street,
-        bairro: data.neighborhood,
-        cidade: data.city,
-        estado: data.state,
-      }))
-    }
-    if (form.cep.replace(/\D/g, '').length === 8) lookup()
-  }, [form.cep, showError])
-
   const handleChange = (
     e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>,
   ) => {
     const { name, value } = e.target
     setForm((prev) => ({ ...prev, [name]: value }))
-    if (name === 'cpf' || name === 'email') {
-      setFieldErrors((prev) => ({ ...prev, [name]: undefined }))
-    }
   }
 
   const handleSelectProduto = (id: string) => {
     setForm((prev) => ({ ...prev, produtoId: id, tamanho: '' }))
   }
 
-  const handleStepValidate = async (index: number) => {
-    if (index === 0) {
-      setFieldErrors({})
-      const cpfNumerico = form.cpf.replace(/\D/g, '')
-      if (!isValidCPF(cpfNumerico)) {
-        setFieldErrors({ cpf: 'CPF inválido.' })
-        return false
-      }
-      if (!isValidEmail(form.email)) {
-        setFieldErrors({ email: 'E-mail inválido.' })
-        return false
-      }
-      try {
-        const query =
-          `/api/usuarios/exists?cpf=${cpfNumerico}&email=${encodeURIComponent(form.email)}${
-            isLoggedIn && user ? `&excludeId=${user.id}` : ''
-          }`
-        const res = await fetch(query)
-        if (res.ok) {
-          const data = await res.json()
-          if (data.cpf) {
-            setFieldErrors((prev) => ({ ...prev, cpf: 'CPF já cadastrado.' }))
-            return false
-          }
-          if (data.email) {
-            setFieldErrors((prev) => ({ ...prev, email: 'E-mail já cadastrado.' }))
-            return false
-          }
-        }
-      } catch {
-        /* ignore */
-      }
-    }
-    return true
-  }
-
   const handleSubmit = async () => {
     setLoading(true)
     try {
-      if (!liderId && !isLoggedIn && form.password.length < 8) {
-        showError('A senha deve ter ao menos 8 caracteres.')
-        setLoading(false)
-        return
-      }
-      if (!liderId && !isLoggedIn && form.password !== form.passwordConfirm) {
-        showError('As senhas não coincidem.')
+      const userData = user
+      if (!userData) {
+        showError('Usuário não autenticado.')
         setLoading(false)
         return
       }
       const url = liderId ? '/api/inscricoes' : '/loja/api/inscricoes'
-      const [firstName, ...rest] = form.nome.split(' ')
-      const payload = liderId
-        ? {
-            nome: form.nome,
-            email: form.email,
-            telefone: form.telefone,
-            cpf: form.cpf,
-            data_nascimento: form.data_nascimento,
-            genero: form.genero,
-            cep: form.cep,
-            endereco: form.endereco,
-            bairro: form.bairro,
-            estado: form.estado,
-            cidade: form.cidade,
-            numero: form.numero,
-            role: form.role,
-            produtoId: form.produtoId,
-            tamanho: form.tamanho,
-            liderId,
-            eventoId,
-            paymentMethod: form.paymentMethod,
-          }
-        : {
-            user_first_name: firstName,
-            user_last_name: rest.join(' '),
-            user_email: form.email,
-            user_phone: form.telefone,
-            user_cpf: form.cpf,
-            user_birth_date: form.data_nascimento,
-            user_gender: form.genero,
-            user_cep: form.cep,
-            user_address: form.endereco,
-            user_neighborhood: form.bairro,
-            user_state: form.estado,
-            user_city: form.cidade,
-            user_number: form.numero,
-            campo: form.campoId,
-            evento: eventoId,
-            produtoId: form.produtoId,
-            tamanho: form.tamanho,
-            ...(isLoggedIn
-              ? {}
-              : {
-                  password: form.password,
-                  passwordConfirm: form.passwordConfirm,
-                  role: form.role,
-                }),
-            paymentMethod: form.paymentMethod,
-          }
+      const [firstName, ...rest] = String(userData.nome || '').split(' ')
+      const base = {
+        user_first_name: firstName,
+        user_last_name: rest.join(' '),
+        user_email: userData.email,
+        user_phone: userData.telefone,
+        user_cpf: userData.cpf,
+        user_birth_date: String(userData.data_nascimento ?? '').split(' ')[0],
+        user_gender: form.genero || userData.genero,
+        user_cep: userData.cep,
+        user_address: userData.endereco,
+        user_neighborhood: userData.bairro,
+        user_state: userData.estado,
+        user_city: userData.cidade,
+        user_number: userData.numero,
+        campo: form.campoId,
+        evento: eventoId,
+        produtoId: form.produtoId,
+        tamanho: form.tamanho,
+        paymentMethod: form.paymentMethod,
+      }
+      const payload = liderId ? { ...base, liderId } : base
       const res = await fetch(url, {
         method: 'POST',
         headers: { ...getAuthHeaders(pb), 'Content-Type': 'application/json' },
@@ -372,61 +236,32 @@ export default function EventForm({ eventoId, liderId }: EventFormProps) {
       setLoading(false)
     }
   }
+  const handleStepValidate = async (index: number) => {
+    if (!signupDone && index === 0) {
+      const ok = await createUserRef.current?.submit()
+      return Boolean(ok)
+    }
+    return true
+  }
 
-  const steps = [
-    {
-      title: 'Dados Pessoais',
+  const steps: { title: string; content: React.ReactNode }[] = []
+
+  if (!signupDone) {
+    steps.push({
+      title: 'Criar Conta',
+      content: (
+        <CreateUserForm
+          ref={createUserRef}
+          onSuccess={() => setSignupDone(true)}
+          showButton={false}
+        />
+      ),
+    })
+  } else {
+    steps.push({
+      title: "Informações Adicionais",
       content: (
         <div className="space-y-4">
-          <FormField label="Nome" htmlFor="nome">
-            <TextField
-              id="nome"
-              name="nome"
-              value={form.nome}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="E-mail" htmlFor="email" error={fieldErrors.email}>
-            <TextField
-              id="email"
-              name="email"
-              type="email"
-              value={form.email}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="Telefone" htmlFor="telefone">
-            <InputWithMask
-              id="telefone"
-              name="telefone"
-              mask="telefone"
-              value={form.telefone}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="CPF" htmlFor="cpf" error={fieldErrors.cpf}>
-            <InputWithMask
-              id="cpf"
-              name="cpf"
-              mask="cpf"
-              value={form.cpf}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="Data de Nascimento" htmlFor="data_nascimento">
-            <TextField
-              id="data_nascimento"
-              name="data_nascimento"
-              type="date"
-              value={form.data_nascimento}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
           <FormField label="Gênero" htmlFor="genero">
             <select
               id="genero"
@@ -443,69 +278,10 @@ export default function EventForm({ eventoId, liderId }: EventFormProps) {
           </FormField>
         </div>
       ),
-    },
-    {
-      title: 'Endereço',
-      content: (
-        <div className="space-y-4">
-          <FormField label="CEP" htmlFor="cep">
-            <TextField
-              id="cep"
-              name="cep"
-              value={form.cep}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="Endereço" htmlFor="endereco">
-            <TextField
-              id="endereco"
-              name="endereco"
-              value={form.endereco}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="Estado" htmlFor="estado">
-            <TextField
-              id="estado"
-              name="estado"
-              value={form.estado}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="Cidade" htmlFor="cidade">
-            <TextField
-              id="cidade"
-              name="cidade"
-              value={form.cidade}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="Bairro" htmlFor="bairro">
-            <TextField
-              id="bairro"
-              name="bairro"
-              value={form.bairro}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="Número" htmlFor="numero">
-            <TextField
-              id="numero"
-              name="numero"
-              value={form.numero}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-        </div>
-      ),
-    },
-  ]
+    })
+  }
+
+
 
   if (liderId) {
     steps.push({
@@ -641,35 +417,6 @@ export default function EventForm({ eventoId, liderId }: EventFormProps) {
               <option value="boleto">Boleto</option>
             </select>
           </FormField>
-          
-        </div>
-      ),
-    })
-  }
-
-  if (!liderId && !isLoggedIn) {
-    steps.push({
-      title: 'Criar Senha',
-      content: (
-        <div className="space-y-4">
-          <FormField label="Senha" htmlFor="password">
-            <PasswordField
-              id="password"
-              name="password"
-              value={form.password}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
-          <FormField label="Confirme a Senha" htmlFor="passwordConfirm">
-            <PasswordField
-              id="passwordConfirm"
-              name="passwordConfirm"
-              value={form.passwordConfirm}
-              onChange={handleChange}
-              required
-            />
-          </FormField>
         </div>
       ),
     })
@@ -705,8 +452,8 @@ export default function EventForm({ eventoId, liderId }: EventFormProps) {
     <FormWizard
       steps={steps}
       onFinish={handleSubmit}
-      onStepValidate={handleStepValidate}
       loading={loading}
+      onStepValidate={handleStepValidate}
       className="max-w-lg mx-auto bg-white p-6 rounded-xl shadow"
     />
   )
