+++ components/templates/index.ts
@@ -7,3 +7,4 @@ export { default as HeaderLoja } from './HeaderLoja'
 export { default as FooterLoja } from './FooterLoja'
 export { default as LoginForm } from './LoginForm'
 export { default as SignUpForm } from './SignUpForm'
+export { default as CreateUserForm } from './CreateUserForm'
