+++ stories/SignUpForm.stories.tsx
@@ -1,12 +1,12 @@
 import type { Meta, StoryObj } from '@storybook/nextjs'
 import { within, expect } from 'storybook/test'
-import SignUpForm from '@/components/templates/SignUpForm'
+import CreateUserForm from '@/components/templates/CreateUserForm'
 import { AuthProvider } from '../lib/context/AuthContext'
 import { ThemeProvider } from '../lib/context/ThemeContext'
 
 const meta = {
-  title: 'Components/SignUpForm',
-  component: SignUpForm,
+  title: 'Components/CreateUserForm',
+  component: CreateUserForm,
   decorators: [
     (Story) => (
       <ThemeProvider>
@@ -17,7 +17,7 @@ const meta = {
     ),
   ],
   tags: ['autodocs'],
-} satisfies Meta<typeof SignUpForm>
+} satisfies Meta<typeof CreateUserForm>
 
 export default meta
